// utils/types/types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
  Long: any;
  Time: any;
};








export type DivisionType = {
  __typename?: 'DivisionType';
  divisionId?: Maybe<Scalars['String']>;
  _id: Scalars['ID'];
  teams: TeamTypePage;
  country?: Maybe<Scalars['String']>;
  divisionName?: Maybe<Scalars['String']>;
  events: EventPage;
  _ts: Scalars['Long'];
};
// ...
export type DivisionTypePage = {
  __typename?: 'DivisionTypePage';
  data: Array<Maybe<DivisionType>>;
  after?: Maybe<Scalars['String']>;
  before?: Maybe<Scalars['String']>;
};

export type DivisionTypeTeamsRelation = {
  create?: Maybe<Array<Maybe<TeamTypeInput>>>;
  connect?: Maybe<Array<Maybe<Scalars['ID']>>>;
  disconnect?: Maybe<Array<Maybe<Scalars['ID']>>>;
};

export type Event = {
  __typename?: 'Event';
  divisionId?: Maybe<DivisionType>;
  awayTeamName?: Maybe<Scalars['String']>;
  lastUpdate?: Maybe<Scalars['String']>;
  _id: Scalars['ID'];
  finished?: Maybe<Scalars['Boolean']>;
  homeTeamId?: Maybe<TeamType>;
  dateTime?: Maybe<Scalars['String']>;
  eventId?: Maybe<Scalars['String']>;
  divisionName?: Maybe<Scalars['String']>;
  submittedGuesses: UserGuessPage;
  awayTeamId?: Maybe<TeamType>;
  homeTeamName?: Maybe<Scalars['String']>;
  venue?: Maybe<Scalars['String']>;
  _ts: Scalars['Long'];
};
// ...
export type EventPage = {
  __typename?: 'EventPage';
  data: Array<Maybe<Event>>;
  after?: Maybe<Scalars['String']>;
  before?: Maybe<Scalars['String']>;
};

export type EventSubmittedGuessesRelation = {
  create?: Maybe<Array<Maybe<UserGuessInput>>>;
  connect?: Maybe<Array<Maybe<Scalars['ID']>>>;
  disconnect?: Maybe<Array<Maybe<Scalars['ID']>>>;
};

export type League = {
  __typename?: 'League';
  standings: StandingsPage;
  name: Scalars['String'];
  _id: Scalars['ID'];
  slug: Scalars['String'];
  options: LeagueOptions;
  members: UserPage;
  _ts: Scalars['Long'];
};
// ...
export type LeagueOptions = {
  __typename?: 'LeagueOptions';
  _id: Scalars['ID'];
  teams?: Maybe<Array<Maybe<TeamType>>>;
  public?: Maybe<Scalars['Boolean']>;
  divisions?: Maybe<Array<Maybe<Scalars['String']>>>;
  class?: Maybe<LeagueType>;
  _ts: Scalars['Long'];
};

export type LeagueOptionsInput = {
  class?: Maybe<LeagueType>;
  divisions?: Maybe<Array<Maybe<Scalars['String']>>>;
  teams?: Maybe<Array<Maybe<Scalars['ID']>>>;
  public?: Maybe<Scalars['Boolean']>;
};

export type LeagueOptionsRelation = {
  create?: Maybe<LeagueOptionsInput>;
  connect?: Maybe<Scalars['ID']>;
};

export type LeaguePage = {
  __typename?: 'LeaguePage';
  data: Array<Maybe<League>>;
  after?: Maybe<Scalars['String']>;
  before?: Maybe<Scalars['String']>;
};
// ...
export type StandingsPage = {
  __typename?: 'StandingsPage';
  data: Array<Maybe<Standings>>;
  after?: Maybe<Scalars['String']>;
  before?: Maybe<Scalars['String']>;
};

export type TeamType = {
  __typename?: 'TeamType';
  teamName?: Maybe<Scalars['String']>;
  teamId?: Maybe<Scalars['String']>;
  _id: Scalars['ID'];
  homeEvents: EventPage;
  country?: Maybe<Scalars['String']>;
  division: DivisionTypePage;
  inLeagues?: Maybe<Array<Maybe<LeagueOptions>>>;
  awayEvents: EventPage;
  badge?: Maybe<Scalars['String']>;
  _ts: Scalars['Long'];
};
// ...
export type TeamTypePage = {
  __typename?: 'TeamTypePage';
  data: Array<Maybe<TeamType>>;
  after?: Maybe<Scalars['String']>;
  before?: Maybe<Scalars['String']>;
};


export type User = {
  __typename?: 'User';
  leagues: LeaguePage;
  email: Scalars['String'];
  role?: Maybe<UserRole>;
  username: Scalars['String'];
  _id: Scalars['ID'];
  guesses?: Maybe<Array<Maybe<UserGuess>>>;
  favTeam?: Maybe<Scalars['String']>;
  _ts: Scalars['Long'];
};


export type UserLeaguesArgs = {
  _size?: Maybe<Scalars['Int']>;
  _cursor?: Maybe<Scalars['String']>;
};

export type UserGuess = {
  __typename?: 'UserGuess';
  winner?: Maybe<WinnerEnum>;
  _id: Scalars['ID'];
  league?: Maybe<League>;
  score?: Maybe<Array<Maybe<Scalars['String']>>>;
  eventId?: Maybe<Event>;
  apiEventId?: Maybe<Scalars['String']>;
	corrected?: Maybe<Scalars['Boolean']>;
  user?: Maybe<User>;
  _ts: Scalars['Long'];
};
// ...
export type UserGuessPage = {
  __typename?: 'UserGuessPage';
  data: Array<Maybe<UserGuess>>;
  after?: Maybe<Scalars['String']>;
  before?: Maybe<Scalars['String']>;
};
// ...
export type UserPage = {
  __typename?: 'UserPage';
  data: Array<Maybe<User>>;
  after?: Maybe<Scalars['String']>;
  before?: Maybe<Scalars['String']>;
};

export enum UserRole {
  Admin = 'ADMIN',
  Guest = 'GUEST',
  Member = 'MEMBER'
}

export enum WinnerEnum {
  Homewin = 'HOMEWIN',
  Draw = 'DRAW',
  Awaywin = 'AWAYWIN'
}